+++ CHANGELOG.md
@@ -18,6 +18,14 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
   - `waystonewarps.bypass.rename`: Allows access to rename the waystone.
   - `waystonewarps.bypass.icon`: Allows access to change the waystone icon.
   - `waystonewarps.bypass.relocate`: Allows access to relocate the waystone.
+- Add new command system to remove invalid warp world data:
+  - `/waystonewarps invalids list`: Lists missing worlds containing warps
+  - `/waystonewarps invalids remove <id>` Removes warps for a given world.
+  - `/waystonewarps invalids removeall`: Removes all warps for missing worlds.
+- New permissions to use the invalids command system:
+  - `waystonewarps.admin.invalids.list`: Allows usage of the list command
+  - `waystonewarps.admin.invalids.remove`: Allows usage of remove command
+  - `waystonewarps.admin.invalids.removeall`: Allows usage of removeall command
 
 ### Changed
 - Menu buttons are now gold with a grey description for standardisation.
@@ -25,6 +33,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 ### Fixed
 - Config not being read on the first launch, required restart before the plugin could actually let you do anything.
 - Waystones not being covered by protection plugins even though the block break itself is prevented.
+- Whitelists not being removed from the database when the waystone is removed.
 
 ## [0.3.5]